Command formatting to be more robust, based on the arguments that IO.popen can take.
Commands with spaces, such as "bundle exec rake resque:work" would fail due to the way the command was being passed into IO.popen. After looking through the IO.popen documentation it turns out it can be passed an array of commands, and also has the ability to take a hash as the environmental variables in the first position of that array. This fixes that issue, and the test suite is passing.

I've also changed the retrieval of the @command variable from the options hash to use .fetch. This will change the behavior so that, the initialize method will fail if the :command key does not exist on the options hash. The missing key exception is much more intuitive than an exception stating that .split was called on nil(which would be the case if the command was retrieved using the square brackets and it did not exist)